git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@6114 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ MagickCore/transform.c
@@ -747,8 +747,8 @@ MagickExport Image *CropImageToTiles(const Image *image,
           width+=(geometry.x < 0 ? -1 : 1)*geometry.x;
           height+=(geometry.y < 0 ? -1 : 1)*geometry.y;
         }
-      delta.x=((double)width)/((double)geometry.width);
-      delta.y=((double)height)/((double)geometry.height);
+      delta.x=(double) width/geometry.width;
+      delta.y=(double) height/geometry.height;
       for (offset.y=0; offset.y < (double) height; )
       {
         if ((flags & AspectValue) == 0)
